gpu: nvgpu: obj_ctx: fix unsigned int cast cert error
Fix CERT-C error for translating size from "unsigned long" to
"unsigned int".

Error: CERT INT31-C:
nvgpu/drivers/gpu/nvgpu/common/gr/obj_ctx.c:300:
cert_violation: Casting "size" from "unsigned long" to "unsigned int"
 without checking its value may result in lost or misinterpreted data.

JIRA NVGPU-3409

Reviewed-on: https://git-master.nvidia.com/r/2114536
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/gr/obj_ctx.c
@@ -284,7 +284,7 @@ int nvgpu_gr_obj_ctx_commit_global_ctx_buffers(struct gk20a *g,
 	struct nvgpu_gr_config *config,	struct nvgpu_gr_ctx *gr_ctx, bool patch)
 {
 	u64 addr;
-	u32 size;
+	size_t size;
 
 	nvgpu_log_fn(g, " ");
 
@@ -298,7 +298,7 @@ int nvgpu_gr_obj_ctx_commit_global_ctx_buffers(struct gk20a *g,
 
 	/* global pagepool buffer */
 	addr = nvgpu_gr_ctx_get_global_ctx_va(gr_ctx, NVGPU_GR_CTX_PAGEPOOL_VA);
-	size = (u32)nvgpu_gr_global_ctx_get_size(global_ctx_buffer,
+	size = nvgpu_gr_global_ctx_get_size(global_ctx_buffer,
 			NVGPU_GR_GLOBAL_CTX_PAGEPOOL);
 
 	g->ops.gr.init.commit_global_pagepool(g, gr_ctx, addr, size, patch,
