+++ drivers/gpu/nvgpu/hal/gr/init/gr_init_gm20b.h
@@ -85,7 +85,7 @@ void gm20b_gr_init_commit_global_bundle_cb(struct gk20a *g,
 	struct nvgpu_gr_ctx *gr_ctx, u64 addr, u64 size, bool patch);
 u32 gm20b_gr_init_pagepool_default_size(struct gk20a *g);
 void gm20b_gr_init_commit_global_pagepool(struct gk20a *g,
-	struct nvgpu_gr_ctx *gr_ctx, u64 addr, u32 size, bool patch,
+	struct nvgpu_gr_ctx *gr_ctx, u64 addr, size_t size, bool patch,
 	bool global_ctx);
 void gm20b_gr_init_commit_global_attrib_cb(struct gk20a *g,
 	struct nvgpu_gr_ctx *gr_ctx, u32 tpc_count, u32 max_tpc, u64 addr,
